Move CNAME tests to upload suite
